Refactor Links exporter to allow different subclasses
- Split out existing functions into LocalAuthorityLinksExporter
