+++ spec/lib/local-links-manager/export/fixtures/ni_link.csv
@@ -1,2 +1,2 @@
-Authority Name,SNAC,GSS,Description,LGSL,LGIL,URL,Supported by GOV.UK
-Belfast,,456,Service 123: Interaction 1,123,1,http://www.example.com/456/1,true
+Authority Name,GSS,Description,LGSL,LGIL,URL,Supported by GOV.UK
+Belfast,456,Service 123: Interaction 1,123,1,http://www.example.com/456/1,true
